+++ gulpfile.js
@@ -1,12 +1,16 @@
 var gulp = require('gulp');
 var phpunit = require('gulp-phpunit');
+var plumber = require('gulp-plumber');
 
 gulp.task('phpunit', function() {
-    gulp.src('phpunit.xml').pipe(phpunit('phpunit'));
+    gulp.src('')
+        .pipe(plumber())
+        .pipe(phpunit('phpunit'))
+        .pipe(plumber.stop());
 });
 
 gulp.task('watch', function () {
-    gulp.watch(['src/*.php','tests/*.php'], ['phpunit']);
+    gulp.watch(['**/*.php'], ['phpunit']);
 });
 
 // What tasks does running gulp trigger?