+++ src/LocalStripeServiceProvider.php
@@ -6,19 +6,34 @@ use Illuminate\Support\ServiceProvider;
 class LocalStripeServiceProvider extends ServiceProvider 
 {
 
-	protected $defer = true;    // only load if/when needed
+    protected $defer = true;    // only load if/when needed
 
-	public function register()
-	{
-		$this->app['local-stripe'] = $this->app->share(function($app){
-			return new LocalStripe();
-		});
-	}
+    public function boot()
+    {
+        $this->publishes([
+            realpath(__DIR__.'/src/database/migrations') => $this->app->databasePath().'/migrations',
+        ]);
+    }
 
-	public function provides()
-	{
-		return array('LocalStripe');
-	}
+    public function register()
+    {
+        $this->app['local-stripe'] = $this->app->share(function($app){
+
+            $connector = new Connector;
+            $connector->setApiKey(getenv('STRIPE_SECRET'));
+
+            $parser = new ParameterParser;
+            $pusher = new Pusher($connector, $parser);
+            $fetcher = new Fetcher($connector);
+
+            return new LocalStripe($connector, $pusher, $fetcher);
+        });
+    }
+
+    public function provides()
+    {
+        return array('LocalStripe');
+    }
 
 }